extract InvitationStatus into separate components
